Move util to api module
