+++ api/src/main/java/gov/cms/ab2d/api/controller/ErrorHandler.java
@@ -10,7 +10,7 @@ import gov.cms.ab2d.api.security.InvalidAuthHeaderException;
 import gov.cms.ab2d.api.security.MissingTokenException;
 import gov.cms.ab2d.common.service.InvalidClientInputException;
 import gov.cms.ab2d.common.service.InvalidContractException;
-import gov.cms.ab2d.common.util.UtilMethods;
+import gov.cms.ab2d.api.util.UtilMethods;
 import gov.cms.ab2d.eventclient.clients.SQSEventClient;
 import gov.cms.ab2d.eventclient.config.Ab2dEnvironment;
 import gov.cms.ab2d.eventclient.events.ApiResponseEvent;
